refactor: split profiles.go into helpers and fix warnings / linting errors
